ARROW-7610: [Java] Finish support for 64 bit int allocations
Modified the `NettyAllocationManager` to allow for 64 bit allocations. If the requested size is > `Int.MAX_VALUE`, memory will be allocated using `PlatformDependent.allocateMemory(long)` otherwise the existing allocation method remains unchanged.

Unit tests were added that uses a custom AllocationManager with a smaller cutoff value to be able to test the new functionality within our current CI memory limits. Additional tests were added that allocate large buffers, but are required to run manually.

Closes #6323 from liyafan82/fly_0123_alloc

Authored-by: liyafan82 <fan_li_ya@foxmail.com>

+++ java/vector/src/main/java/org/apache/arrow/vector/DateMilliVector.java
@@ -116,7 +116,7 @@ public final class DateMilliVector extends BaseFixedWidthVector {
     if (NULL_CHECKING_ENABLED && isSet(index) == 0) {
       throw new IllegalStateException("Value at index is null");
     }
-    return valueBuffer.getLong(index * TYPE_WIDTH);
+    return valueBuffer.getLong((long) index * TYPE_WIDTH);
   }
 
   /**
@@ -132,7 +132,7 @@ public final class DateMilliVector extends BaseFixedWidthVector {
       return;
     }
     holder.isSet = 1;
-    holder.value = valueBuffer.getLong(index * TYPE_WIDTH);
+    holder.value = valueBuffer.getLong((long) index * TYPE_WIDTH);
   }
 
   /**
@@ -145,7 +145,7 @@ public final class DateMilliVector extends BaseFixedWidthVector {
     if (isSet(index) == 0) {
       return null;
     } else {
-      final long millis = valueBuffer.getLong(index * TYPE_WIDTH);
+      final long millis = valueBuffer.getLong((long) index * TYPE_WIDTH);
       return DateUtility.getLocalDateTimeFromEpochMilli(millis);
     }
   }
@@ -158,7 +158,7 @@ public final class DateMilliVector extends BaseFixedWidthVector {
 
 
   private void setValue(int index, long value) {
-    valueBuffer.setLong(index * TYPE_WIDTH, value);
+    valueBuffer.setLong((long) index * TYPE_WIDTH, value);
   }
 
   /**
@@ -282,7 +282,7 @@ public final class DateMilliVector extends BaseFixedWidthVector {
    * @return value stored at the index.
    */
   public static long get(final ArrowBuf buffer, final int index) {
-    return buffer.getLong(index * TYPE_WIDTH);
+    return buffer.getLong((long) index * TYPE_WIDTH);
   }
