+++ java/vector/src/main/java/org/apache/arrow/vector/Float4Vector.java
@@ -114,7 +114,7 @@ public final class Float4Vector extends BaseFixedWidthVector implements Floating
     if (NULL_CHECKING_ENABLED && isSet(index) == 0) {
       throw new IllegalStateException("Value at index is null");
     }
-    return valueBuffer.getFloat(index * TYPE_WIDTH);
+    return valueBuffer.getFloat((long) index * TYPE_WIDTH);
   }
 
   /**
@@ -130,7 +130,7 @@ public final class Float4Vector extends BaseFixedWidthVector implements Floating
       return;
     }
     holder.isSet = 1;
-    holder.value = valueBuffer.getFloat(index * TYPE_WIDTH);
+    holder.value = valueBuffer.getFloat((long) index * TYPE_WIDTH);
   }
 
   /**
@@ -143,7 +143,7 @@ public final class Float4Vector extends BaseFixedWidthVector implements Floating
     if (isSet(index) == 0) {
       return null;
     } else {
-      return valueBuffer.getFloat(index * TYPE_WIDTH);
+      return valueBuffer.getFloat((long) index * TYPE_WIDTH);
     }
   }
 
@@ -155,7 +155,7 @@ public final class Float4Vector extends BaseFixedWidthVector implements Floating
 
 
   private void setValue(int index, float value) {
-    valueBuffer.setFloat(index * TYPE_WIDTH, value);
+    valueBuffer.setFloat((long) index * TYPE_WIDTH, value);
   }
 
   /**
@@ -279,7 +279,7 @@ public final class Float4Vector extends BaseFixedWidthVector implements Floating
    * @return value stored at the index.
    */
   public static float get(final ArrowBuf buffer, final int index) {
-    return buffer.getFloat(index * TYPE_WIDTH);
+    return buffer.getFloat((long) index * TYPE_WIDTH);
   }
 
   @Override
