+++ java/vector/src/main/java/org/apache/arrow/vector/TimeStampSecVector.java
@@ -116,7 +116,7 @@ public final class TimeStampSecVector extends TimeStampVector {
       return;
     }
     holder.isSet = 1;
-    holder.value = valueBuffer.getLong(index * TYPE_WIDTH);
+    holder.value = valueBuffer.getLong((long) index * TYPE_WIDTH);
   }
 
   /**
@@ -129,7 +129,7 @@ public final class TimeStampSecVector extends TimeStampVector {
     if (isSet(index) == 0) {
       return null;
     } else {
-      final long secs = valueBuffer.getLong(index * TYPE_WIDTH);
+      final long secs = valueBuffer.getLong((long) index * TYPE_WIDTH);
       final long millis = java.util.concurrent.TimeUnit.SECONDS.toMillis(secs);
       return DateUtility.getLocalDateTimeFromEpochMilli(millis);
     }
